+++ modules/tinymce/src/core/test/ts/browser/html/DomParserTest.ts
@@ -1695,12 +1695,20 @@ describe('browser.tinymce.core.html.DomParserTest', () => {
       assert.equal(serializedHtml, '<svg></svg><p>foo</p>');
     });
 
-    it('TINY-10237: Should retain SVG elements as is but filter out scripts', () => {
+    it('TINY-10237: Should retain SVG elements as is but filter out scripts and invalid children', () => {
       const schema = Schema();
       schema.addValidElements('svg[*]');
       const input = '<svg><circle><desc><b>foo</b><script>alert(1)</script></desc></circle></svg>foo';
       const serializedHtml = HtmlSerializer({}, schema).serialize(DomParser({ forced_root_block: 'p' }, schema).parse(input));
-      assert.equal(serializedHtml, '<svg><circle><desc><b>foo</b></desc></circle></svg><p>foo</p>');
+      assert.equal(serializedHtml, '<svg><circle><desc></desc></circle></svg><p>foo</p>');
+    });
+
+    it('TINY-11331: Should retain SVG elements and keep HTML elements that are valid inside an SVG', () => {
+      const schema = Schema();
+      schema.addValidElements('svg[*]');
+      const input = '<svg><a href="/docs/Web/SVG/Element/circle"><circle cx="50" cy="40" r="35" /></a><script>alert(1)</script></svg>foo';
+      const serializedHtml = HtmlSerializer({}, schema).serialize(DomParser({ forced_root_block: 'p' }, schema).parse(input));
+      assert.equal(serializedHtml, '<svg><a href="/docs/Web/SVG/Element/circle"><circle cx="50" cy="40" r="35"></circle></a></svg><p>foo</p>');
     });
 
     it('TINY-10237: Should retain SVG elements and keep scripts if sanitize is set to false', () => {