+++ modules/tinymce/src/core/test/ts/browser/html/SanitizationTest.ts
@@ -55,7 +55,7 @@ describe('browser.tinymce.core.html.SanitizationTest', () => {
 
     it('Sanitize SVG with mixed HTML', () => testNamespaceSanitizer({
       input: '<svg><circle cx="50" cy="50" r="40" stroke="black" stroke-width="3" fill="red"><desc><script>alert(1)</script><p>hello</p></circle></a></svg>',
-      expected: '<svg><circle cx="50" cy="50" r="40" stroke="black" stroke-width="3" fill="red"><desc><p>hello</p></desc></circle></svg>'
+      expected: '<svg><circle cx="50" cy="50" r="40" stroke="black" stroke-width="3" fill="red"><desc></desc></circle></svg>'
     }));
 
     it('Sanitize SVG with xlink with script url', () => testNamespaceSanitizer({